Untracked files: 	static/Fonts/ 	static/Javascript/ 	static/favicon.ico 	templates/

+++ server.sh
@@ -1,9 +1,15 @@
 #!/bin/bash
 #d
 clear
+if [ -e "flask.db" ]
+then
 echo "[+] Removing old DB"
 rm flask.db
+fi
+if [ -e "log.log" ]
+then
 rm log.log
+fi
 echo "[+] Initializing DB"
 python start.py
 echo "[+] Starting server"
